Moved the plugin/command code to its own module
